move view state transforms into their own folder
