+++ features/borrow/manage/pipes/viewStateTransforms/manageVaultAllowances.ts
@@ -1,7 +1,7 @@
 import { BigNumber } from 'bignumber.js'
 import { maxUint256 } from 'blockchain/calls/erc20'
 
-import { ManageStandardBorrowVaultState,ManageVaultChange } from './manageVault'
+import { ManageStandardBorrowVaultState, ManageVaultChange } from '../manageVault'
 
 export const allowanceDefaults: Partial<ManageStandardBorrowVaultState> = {
   collateralAllowanceAmount: maxUint256,
