+++ features/borrow/manage/pipes/viewStateTransforms/manageVaultCalculations.ts
@@ -4,7 +4,7 @@ import { Vault } from 'blockchain/vaults'
 import { BalanceInfo } from 'features/shared/balanceInfo'
 import { one, zero } from 'helpers/zero'
 
-import { ManageStandardBorrowVaultState } from './manageVault'
+import { ManageStandardBorrowVaultState } from '../manageVault'
 
 // This value ought to be coupled in relation to how much we round the raw debt
 // value in the vault (vault.debt)
