+++ features/borrow/manage/pipes/viewStateTransforms/manageVaultForm.ts
@@ -1,4 +1,4 @@
-import { MainAction, ManageStandardBorrowVaultState,ManageVaultChange } from './manageVault'
+import { MainAction, ManageStandardBorrowVaultState, ManageVaultChange } from '../manageVault'
 import { allowanceDefaults } from './manageVaultAllowances'
 import { depositAndGenerateDefaults, paybackAndWithdrawDefaults } from './manageVaultInput'
