+++ features/borrow/manage/pipes/viewStateTransforms/manageVaultTransactions.ts
@@ -14,8 +14,8 @@ import { zero } from 'helpers/zero'
 import { iif, Observable, of } from 'rxjs'
 import { filter, first, switchMap } from 'rxjs/operators'
 
-import { TxError } from '../../../../helpers/types'
-import { ManageStandardBorrowVaultState,ManageVaultChange } from './manageVault'
+import { TxError } from '../../../../../helpers/types'
+import { ManageStandardBorrowVaultState, ManageVaultChange } from '../manageVault'
 
 type ProxyChange =
   | {
